finish phase2

+++ myshop/order/models.py
@@ -8,7 +8,8 @@ class Order(models.Model):
     customer = models.ForeignKey(Customer, on_delete=models.CASCADE)
     order_number = models.CharField(max_length=20)
     order_date = models.DateTimeField(auto_now=True)
-    total_price = models.IntegerField(blank=True, null=True)
+    total_price = models.DecimalField(
+        decimal_places=2, max_digits=11, blank=True, null=True)
     tax = models.IntegerField(blank=True, null=True)
 
     CANCELED = 'CANC'
@@ -33,8 +34,8 @@ class OrderItem(models.Model):
     order = models.ForeignKey(Order, on_delete=CASCADE)
     product = models.ForeignKey(Product, on_delete=CASCADE)
     quantity = models.IntegerField(default=0)
-    price = models.IntegerField(blank=True, null=True)
-
+    price = models.DecimalField(
+        decimal_places=2, max_digits=11, blank=True, null=True)
     CANCELED = 'CANC'
     PENDING = 'PEND'
     APPROVED = 'APPR'
